Code is merged in index.js. Separate files removed.
