--- LinkedList/ListNode.js
@@ -1,13 +1,0 @@
-class ListNode {
-
-	/**
-	 * @what instantiates a node with data provided
-	 * @param data of the node
-	 */
-	constructor(data) {
-		this.data = data;
-		this.next = null;
-	}
-}
-
-module.exports = ListNode;
